saturn: Cleanup and modernize.
